% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_date_cols.R
\name{fh_date_cols}
\alias{fh_date_cols}
\title{select all date stamps based on a input year}
\usage{
fh_date_cols(df, year)
}
\arguments{
\item{df}{- a data frame or data table}

\item{year}{- a string, this specifies the year of date stamps to select,
i.e. if year = "2016" all date stamps from 2016 are selected and returned as vector}
}
\value{
a formatted NDVI Data table, from an list of NDVI values
}
\description{
select all date stamps based on a input year
}
\details{
this functions returns a vector with the colnames of all time stamps of a df
so you could subset your df based on that vector and
cut out all columns which are date stamps
}
\author{
Eric Krueger
}
