% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_NDVI_DF.R
\name{fh_NDVI_DF}
\alias{fh_NDVI_DF}
\title{function to return a formatted ndvi df}
\usage{
fh_NDVI_DF(inlist, layernames, transform_ndvi)
}
\arguments{
\item{inlist}{- list of NDVI layer, layers list == numbers of extracted points/polygons}

\item{layernames}{- vector of Layernames which are set a colnames, need to be ordered}

\item{transform_ndvi}{- if TRUE the NDVI values are transfromed like: (x- 10000)/10000}
}
\value{
a formatted NDIV Data table, from an list of NDVI values
}
\description{
function to return a formatted ndvi df
}
\details{
this function  returns a formatted ndvi df (data table)
inlist is the output of mean_extract() or result of a raster::extract job (if poly centroids were used)
layernames is a vector with the columnnames (date stamps) of the new ndvi dt
}
\examples{
\dontrun{
test_list <- readRDS("inst/data/test_list.RDS")
test_layer_list <- fh_stat_extract(test_list, "mean", na.rm = T)
fh_NDVI_DF(test_layer_list,layernames = c(1:24),transform_ndvi = F)
}
}
\author{
Eric Krueger
}
