% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_parallel_mask.R
\name{fh_parallel_mask}
\alias{fh_parallel_mask}
\title{function to mask data with polygons - in parts}
\usage{
fh_parallel_mask(no_cores, in_raster, in_geometry, parts)
}
\arguments{
\item{no_cores}{- number of cores to register for the parallel job}

\item{in_raster}{- input raster to be masked}

\item{in_geometry}{- mask}

\item{parts}{- the raster is splitted into parts to speed up processing,
parts defines the number of parts}
}
\description{
function to mask data with polygons - in parts
}
\author{
Gregor Didenko, Eric Krueger
}
