% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_raster_transfrom.R
\name{fh_raster_transform}
\alias{fh_raster_transform}
\title{function to perform a raster::extract on very large data sets}
\usage{
fh_raster_transform(in_layerStack, parallel, no_cores, transform_eq)
}
\arguments{
\item{in_layerStack}{-  a raster brick or stack}

\item{transfrom_eq}{- equation to manipulate the cell values i.e. "(x - 10000) /10000"
x represents the cell value}
}
\value{

}
\description{
function to perform a raster::extract on very large data sets
}
\details{
takes a layer stack/brick an transforms all layers-pixel values
provide a transformation eq like: transform_eq = "(x-10000) / 10000"
}
\examples{
\dontrun{
r <- r2 <- raster::raster(ncol=10, nrow=10)
r <- raster::setValues(r,1:raster::ncell(r))
r2 <- raster::setValues(r2,10)
raster_brick <- raster::brick(r,r2)
fh_raster_transform(in_layerStack = raster_brick,
                 transform_eq = "(x + 10) / 2") # adds 10 to each cell
}
}
\author{
Eric Krueger
}
