% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_stat_extract.R
\name{fh_stat_extract}
\alias{fh_stat_extract}
\title{function to return a layer list for fh_NDVI_DF()}
\usage{
fh_stat_extract(extract_results, stat, ...)
}
\arguments{
\item{stat}{- string to specify the statistical parameter to apply i.e. "mean", "max", "min" ...}

\item{extract_result}{- a nested list, length represents the number of layers
from the raster extract, length of nested lists represents the number of polygons used for extraction}
}
\value{
a list with the selected statistical parameter for each poly in each layer
}
\description{
function to return a layer list for fh_NDVI_DF()
}
\examples{
\dontrun{
test_list <- readRDS("inst/data/test_list.RDS")
test_layer_list <- fh_stat_extract(test_list, "mean", na.rm = T)
fh_NDVI_DF(test_layer_list,layernames = c(1:24),transform_ndvi = F)
}
}
\author{
Eric Krueger
}
