% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_write_rowWise.R
\name{fh_write_rowwise}
\alias{fh_write_rowwise}
\alias{fh_write_rowWise}
\title{function to write a raster file by row}
\usage{
fh_write_rowWise(layer, path, name)
}
\arguments{
\item{layer}{- raster layer to be written}

\item{path}{- file location}

\item{name}{- file name}
}
\value{
writes raster to disk
}
\description{
function to write a raster file by row
}
\details{
write raster to disk rowwise, uses less memory than writeRaster
take raster with path and name and write to disk. good for very large rasters
}
\author{
Gergor Didenko, Eric Krueger
}
